recorder/requirements.txt
afdko==4.0.1
aiofiles==24.1.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
bidict==0.23.1
blinker==1.8.2
booleanOperations==0.9.0
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
defcon==0.10.3
distro==1.9.0
docopt==0.6.2
einops==0.8.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fontMath==0.9.3
fontParts==0.12.1
fontPens==0.2.4
fonttools==4.51.0
frozenlist==1.4.0
fs==2.4.16
fsspec==2024.6.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
Hypercorn==0.17.3
hyperframe==6.0.1
HyperPyYAML==1.2.2
idna==3.7
isbnlib==3.10.14
isbntools==4.3.29
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.3.2
lightning-utilities==0.11.3.post0
llvmlite==0.43.0
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
MutatorMath==3.0.1
netifaces==0.10.6
networkx==3.3
numba==0.60.0
numpy==2.0.0
omegaconf==2.3.0
openai==1.35.10
optuna==3.6.1
packaging==24.1
pandas==2.1.4
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
primePy==1.3
priority==2.0.0
protobuf==5.27.2
pvcobra==2.0.2
pvcobrademo==2.0.2
pvporcupine==3.0.2
pvrecorder==1.2.2
py-translate==1.0.3
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
PyAudio==0.2.14
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytorch-lightning==2.3.2
pytorch-metric-learning==2.5.0
pytz==2023.3.post1
PyYAML==6.0.1
Quart==0.19.6
quart-cors==0.7.0
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.1
scipy==1.14.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
socketio==0.2.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.31
sympy==1.12.1
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==2.3.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.3.1
torchmetrics==1.4.0.post0
tqdm==4.66.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.4
ufonormalizer==0.6.1
ufoProcessor==1.9.0
unicodedata2==15.1.0
urllib3==2.2.2
uvicorn==0.30.1
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.2
zopfli==0.2.3
